+++ src/ts/beatmap/wrapper/eventBox.ts
@@ -13,6 +13,8 @@ export abstract class WrapEventBox<T extends Record<keyof T, unknown>>
     abstract set beatDistribution(value: IWrapEventBox['beatDistribution']);
     abstract get beatDistributionType(): IWrapEventBox['beatDistributionType'];
     abstract set beatDistributionType(value: IWrapEventBox['beatDistributionType']);
+    abstract get easing(): IWrapEventBox['easing'];
+    abstract set easing(value: IWrapEventBox['easing']);
 
     setFilter(value: IWrapIndexFilter) {
         this.filter = value;
@@ -24,6 +26,10 @@ export abstract class WrapEventBox<T extends Record<keyof T, unknown>>
     }
     setBeatDistributionType(value: IWrapEventBox['beatDistributionType']) {
         this.beatDistributionType = value;
+        return this;
+    }
+    setEasing(value: IWrapEventBox['easing']) {
+        this.easing = value;
         return this;
     }