+++ src/ts/types/beatmap/shared/environment.ts
@@ -29,7 +29,8 @@ export type EnvironmentV3Name =
     | 'PyroEnvironment'
     | 'EDMEnvironment'
     | 'TheSecondEnvironment'
-    | 'LizzoEnvironment';
+    | 'LizzoEnvironment'
+    | 'TheWeekndEnvironment';
 /** List of available 360 environment in base game. */
 export type Environment360Name = 'GlassDesertEnvironment';
 /** List of all available environment in base game. */