+++ src/ts/types/beatmap/v3/lightTranslationEventBoxGroup.ts
@@ -1,0 +1,5 @@
+import { IEventBoxGroupTemplate } from './eventBoxGroupTemplate';
+import { ILightTranslationEventBox } from './lightTranslationEventBox';
+
+// deno-lint-ignore no-empty-interface
+export interface ILightTranslationEventBoxGroup extends IEventBoxGroupTemplate<ILightTranslationEventBox> {}