+++ src/ts/types/beatmap/wrapper/lightRotationEventBox.ts
@@ -15,9 +15,10 @@ export interface IWrapLightRotationEventBox extends IWrapEventBox {
      * ```ts
      * 0 -> X
      * 1 -> Y
+     * 2 -> Z
      * ```
      */
-    axis: 0 | 1;
+    axis: 0 | 1 | 2;
     /** Flip rotation `<int>` in light rotation event box. */
     flip: 0 | 1;
     /** Rotation distribution should affect first event `<int>` of light rotation event box. */